refactor(issue-237): mover DiaCalculado a PublicEvents
Unico IPublicEvent del bounded context: el evento que sale hacia otras
aplicaciones del producto. Con esto el Paso 2 queda completo.
